🚚 rename directory
