+++ src/task1/airplane.ts
