+++ src/task1/bird.ts
