+++ src/task1/flyingMuseum.ts
