+++ src/task1/index.ts
