+++ src/task1/penguin.ts
